cleanup and add test
